Start a separate runtime from the compiler, remove logging noise
