+++ lib/getOpalCompilerFilename.js
@@ -45,7 +45,7 @@ function createBundledCompiler(filename) {
     requires = ''
   }
   try {
-    bundlerExecute(`opal --no-exit --no-opal ${flatLoadPaths} -c ${compilerPath} > ${filename}`)
+    bundlerExecute(`opal --no-exit -ghike --no-opal ${flatLoadPaths} -c ${compilerPath} > ${filename}`)
     // need to append the additional requires after we do everything else since we have special monkey patching
     if (requires !== '') {
       bundlerExecute(`opal --no-exit --no-opal -c ${flatLoadPaths} ${requires} >> ${filename}`)
