+++ lib/getOpalRuntimeFilename.js
@@ -1,0 +1,3 @@
+const path = require('path')
+
+module.exports = path.resolve(__dirname, '../vendor/opal_runtime.js')
