+++ lib/resolveFilename.js
@@ -4,7 +4,7 @@ const path = require('path')
 const fs = require('fs')
 
 const opalRequiresForStubbing = require('./getOpalRequiresForStubbing')
-const opalCompilerFilename = require('./getOpalCompilerFilename')()
+const opalCompilerFilename = require('./getOpalRuntimeFilename')
 const useBundler = require('./bundlerCheck')()
 const LOAD_PATH = require('./getRubyMetadata').loadPath
