+++ lib/transpile.js
@@ -9,7 +9,7 @@ const getCompiler = require('./getCompiler')
 const getStub = require('./getStub')
 const getWebpackRequire = require('./getWebpackRequire')
 const getJsRequire = require('./getJsRequire')
-const compilerFilename = require('./getOpalCompilerFilename')()
+const compilerFilename = require('./getOpalRuntimeFilename')
 const opalRequiresForStubbing = require('./getOpalRequiresForStubbing')
 const useBundler = require('./bundlerCheck')()
 const rubyMetadata = require('./getRubyMetadata')
